add filmsupply brand

+++ scripts/fm/db.sh
@@ -1,3 +1,8 @@
+#!/usr/bin/env bash
+
+read -p "Do you need a database import? y/n" -n 1 -r
+echo    # (optional) move to a new line
+if [[ $REPLY =~ ^[Yy]$ ]]; then
 	echo "Adding mysqldump.sql.gz to ~/Homestead"
 	curl -o ~/Homestead/mysqldump.sql.gz https://s3.amazonaws.com/mb-engineering-onboarding/musicbed/mysqldump.sql.gz
 
@@ -5,5 +10,6 @@
 		echo "File not downloaded. Access Denied. Please make sure you are connected to VPN."
 		echo "Cleaning up..."
 		rm -f ~/Homestead/mysqldump.sql.gz
-		[[ "$0" = "$BASH_SOURCE" ]] && exit 1 || return 1 # handle exits from shell or function but don't exit interactive shell
+		exit 1
 	fi
+fi
